Enhance output readability and clarify code comments
The changes enhance readability of the output in the ImmutableMapTest by adding descriptive text to the printed console messages. It also includes an inline comment explaining that the 'toList()' method produces an unmodifiable list from Java 16 onwards.

In VarTypeTest, an unused import is removed, a 'SuppressWarnings' tag "unused" is added to avoid future warnings and a test is added to verify if the var 'list' instance is LinkedList. Moreover, the comment for "getters" has been clarified to reduce ambiguity.