"Refactor Assignment from AstroGateway to AstroResponse"
Removed the Assignment class from AstroGateway and AstroService and moved it to a sub-record in AstroResponse. This was done to improve the modularity of the code and reduce duplications. The use of the Assignment class is related more to the AstroResponse, making it a suitable location for this class. The import statements are updated across the codebase to reflect this change.