Remove preview feature flag and comment out eclipse settings
This commit removes the '--enable-preview' flag in application and test tasks. It also comments out settings specific to the Eclipse IDE. The JVM preview features are no longer needed after the update to the stable version and the Eclipse settings are user-specific and shouldn't be part of the project's common configuration.