Refactor AstroClient and add AstroDemo files
This commit refactors AstroClient class and adds AstroDemo class to the codebase. AstroClient was refactored for better placement of the getAstroResponse and parseJson methods, while a new AstroDemo class is added to provide practical implementation of AstroClient usage. This change was implemented to make the codebase more readable and functional.