Update code in ProcessDictionary classes and add ExampleJShell
The code in the ProcessDictionary and ProcessDictionaryV2 classes has been updated. The method printNTenLongestWords is modified to display the number of longest words in dynamic way and the teeingCollectorDoesBoth method was removed. In the ProcessDictionary class, the object declaration was changed from "var" to "Stream<String>". In addition to these changes, a new class ExampleJShell was added to the project.